+++ Source/OxyPlot/Series/ScatterErrorSeries.cs
@@ -140,6 +140,22 @@ namespace OxyPlot.Series
         }
 
         /// <summary>
+        /// Updates from data fields.
+        /// </summary>
+        protected override void UpdateFromDataFields()
+        {
+            var filler = new ListBuilder<ScatterErrorPoint>();
+            filler.Add(this.DataFieldX, double.NaN);
+            filler.Add(this.DataFieldY, double.NaN);
+            filler.Add(this.DataFieldErrorX, double.NaN);
+            filler.Add(this.DataFieldErrorY, double.NaN);
+            filler.Add(this.DataFieldSize, double.NaN);
+            filler.Add(this.DataFieldValue, double.NaN);
+            filler.Add(this.DataFieldTag, (object)null);
+            filler.FillT(this.ItemsSourcePoints, this.ItemsSource, args => new ScatterErrorPoint(Convert.ToDouble(args[0]), Convert.ToDouble(args[1]), Convert.ToDouble(args[2]), Convert.ToDouble(args[3]), Convert.ToDouble(args[4]), Convert.ToDouble(args[5]), args[6]));
+        }
+
+        /// <summary>
         /// Selects all points for which the passed function returns true.
         /// </summary>
         /// <param name="func">
@@ -151,17 +167,6 @@ namespace OxyPlot.Series
             {
                 this.SelectItem(this.Points.IndexOf(dataPoint));
             }
-        }
-
-        /// <summary>
-        /// Defines the data fields used by the code that reflects on the <see cref="ItemsSeries.ItemsSource" />.
-        /// </summary>
-        /// <param name="filler">The list filler.</param>
-        protected override void DefineDataFields(ListFiller<ScatterErrorPoint> filler)
-        {
-            base.DefineDataFields(filler);
-            filler.Add(this.DataFieldErrorX, (item, value) => item.ErrorX = Convert.ToDouble(value));
-            filler.Add(this.DataFieldErrorY, (item, value) => item.ErrorY = Convert.ToDouble(value));
         }
     }
 }